refactor: extract getXDSChartColors — pure function, hook wraps it
Separates chart color resolution into layers:
- getXDSChartColorsFromResolver(fn) — takes any token resolver, no React
- getXDSChartColors(theme, mode) — takes a defined theme + mode
- useXDSChartColors() — React hook, wraps getXDSChartColorsFromResolver

The hook is now a thin wrapper. Non-React contexts (WebGL, tests, SSR)
use getXDSChartColors directly with a theme object.
