+++ packages/lab/src/Chart/index.ts
@@ -33,8 +33,11 @@ export {
 export {useChart} from './ChartContext';
 export type {ChartContext, ChartMargin, ChartScale} from './types';
 export {m4Reduce, type M4Point} from './m4';
+export {useXDSChartColors} from './useXDSChartColors';
 export {
-  useXDSChartColors,
+  getXDSChartColors,
+  getXDSChartColorsFromResolver,
   type XDSChartColorsAPI,
   type SequentialHue,
-} from './useXDSChartColors';
+  type TokenResolver,
+} from './getXDSChartColors';
